Langchain Basics/requirements.txt
ipykernel
langchain
langchain-core
langchain-groq
langserve
fastapi
uvicorn
python-dotenv
sse_starlette